requirements.txt
Pillow
django
sqlparse
tzdata
asgiref
django-ckeditor